rename diffv2 to diff
