+++ modules/front-end/src/app/core/services/diff-factory.service.ts
@@ -1,7 +1,7 @@
 import {Injectable} from "@angular/core";
 import {RefTypeEnum} from "@core/components/audit-log/types";
-import {FeatureFlagDiffer} from "@shared/diffv2/feature-flag.differ";
-import {IDiffer} from "@shared/diffv2";
+import {FeatureFlagDiffer} from "@shared/diff/feature-flag.differ";
+import {IDiffer} from "@shared/diff";
 
 @Injectable({
   providedIn: 'root'
