+++ modules/front-end/src/app/shared/diff/types.ts
